committing the shop, debugged menu, needs reiew on the cost subtraction and data implemention.
